Prelude: lexing
I'm starting with lexing, and specifically I wanted to try some
data-oriented stuff, so tokens are 4-byte structs which hold their type,
an error flag and an Id (which can be IdentId, NumberId, CharId, etc.)
which points to their underlying data. I initially started doing it the
"normal" way and later changed it so this commit is in-between changes,
I just need to save them to code from mu other computer. Hopefully the
next commit will have a lot more tests, and pass all of them too.