1 - Lexing: Done(?)
Removed the value_store abstraction because while it *could* be useful
I'm not sure it's worth it to take that long to write that shit. Maybe
I'll come back to it later, but for now tokens are stored in simple
vectors, and I upped the Token handle to be usize because memory is
cheap.