CLI flag parsing tests
Added some tests for cli flag parsing. The flags are pretty simple so
the tests are as well.

Note that since the `parse_args()` function checks whether the provided
file exists, we need to pass a file that exists in the test (I chose
"Cargo.toml"). It would be better if we didn't need to do this, maybe
with mocking or something, I don't know. It feels dirty to mock this but
at the same time we're making our tests depend on the environment. Oh
well