Reestructure crates into cargo workspace
The project setup was kind of weird with crates inside crates, it's
better to keep it as a workspace etc. The main crate used to be
ccompiler now it's cli, since it's just meant to be the cli interface.

Also fixed a failing test from the lexer, where '_' was considered a
one-symbol token, which as far as I can tell is wrong, since it's only
valid in identifiers. If it turns out there should actually be a '_'
token I'll add it later.