[Chapter 1][Parser] Initial parsing infra + parse constant
Adds a bunch of changes to lex and ast. Since this is the first time
actually using the `Token`, `Token.handle` and `TokenSource` types as
well all the `NodeKind` internal struct variants most of the changes
were things I realized I needed, such as having a `push` method on
`Tree` which returns the correct node id type. I could go and backport
these changes into the relevant branches/commits but I don't think it's
worth it since I'll probably still be messing with them for some time.

For now the parser has no error handling whatsoever, it panics on
anything it doesn't know how to do. This isn't ideal, but will have to
do for now.