Tests: vendored official tests and added test generator
Nora Sandler makes a bunch of tests available, I thought it would be
nice to turn them into rust native tests so a `cargo test` could just
run them. To do this I want a test-generator that generates simple tests
in rust which read the file, parse it and snapshot assert it, or assert
that it panics. This is a very rough implementation that doesn't work
yet, but the basic idea is there.