[Chapter 1][Assembly AST] Generating assembly ast
This commit introduces the assembly ast, which is another intermediate
representation that more closely resembles assembly. Instead of going
the way of Homogenous nodes stored in a contiguous vector with handles
as references, I just put them inside each other, since this form os ast
is not recursive like the first one: A function in assembly only
contains instructions, and so on. I just didn't want to bother with all
the indexes and so on, they may be better but this way's faster.
This code correctly compiles the simple "int main(void)..." example, but
I haven't tested it further.
I also don't like the crate naming, I think I'll probably change it to
tacky, which is what the book uses.