[Chapter 1][Codegen] initial codegen support
This commit adds initial codegen support. This hasn't been tested with
the compiler suite yet. One thing I just realized was that we generate
assembly files which are not executable yet, this is done by the
official test framework which calls the linker and loader for us. I
would like do also do this, maybe inside the compiler? Not sure