refactor: deduplicate logic in compiler driver stages
Made some changes that make stage functions simpler and less redundant,
since they're basically pipelined instead of having all logic
duplicated. This also makes it easier to test the high level lex
function since it now returns an error. There are still places in the
code where we panic instead of erroring (from parser onwards I think),
but the types are correct, and when I add tests these places should show
up fairly easily.

Also in this commit is a change that makes `Tree` own the
`TokenizedOutput` it references, this is to fix a lifetime issue where
the high-level `parse` function was tokenizing the input inside the
function and then returning a reference to that value, which isn't
allowed since it would be allocated on the stack. Making `Tree` own
`TokenizedOutput` is a simpler mental model and fixes this.