tests: add chapter 1 "valid" integration tests
Added valid program test-cases to native rust. I'm planning to do this
for all chapters since it's just easier and faster. The only downside to
this is that it's testing the whole pipeline, which means that when a
test fails it's harder to figure out which stage's in the wrong. But
this way's pretty painless and I'd rather know whether something's wrong
than have to write a lot more tests or write my own code-generation
thingy (which I already tried) to generate tests for each stage from a
.c file.

I still need to add invalid parsing tests, but I can't yet since our
parser panics instead of returning errors, I need to do that before I
can properly test it.