tests: snapshot assembly string instead of debug repr
I thought that going from our internal assembly representation to a
string would be simple, so it would be ok to leave it out of tests, but
that's not quite right, since there's register allocation, stack
management, and so on. All this involves multiple passes and rewrites of
the ast, which can contain a lot of bugs.